refactor(table-data-grid): move column defs into body
