+++ packages/core/table-data-grid/src/composables/useDatatableColumnDefs.ts
@@ -1,8 +1,11 @@
-import type { TableDataGridHeader } from '../types'
+import type {
+  TableDataGridGridOptions,
+  TableDataGridHeader,
+  TableDataGridRowSelectionMode,
+} from '../types'
 import type {
   TableDataGridColumnDefsConfig,
   TableDataGridColumnDefsGrid,
-  TableDataGridColumnDefsSelection,
   TableDataGridColumnDefsSlots,
 } from '../types/internal'
 import type { ColDef } from 'ag-grid-community'
@@ -10,30 +13,137 @@ import { computed } from 'vue'
 import isEqual from 'lodash-es/isEqual'
 import TableDataGridCellRenderer from '../components/TableDataGridCellRenderer.vue'
 import TableDataGridHeaderRenderer from '../components/TableDataGridHeaderRenderer.vue'
+import TableDataGridSelectionCell from '../components/TableDataGridSelectionCell.vue'
+import TableDataGridSelectionHeader from '../components/TableDataGridSelectionHeader.vue'
 import { isColumnHideable } from '../utils/headers'
+
+const SELECTION_COLUMN_ID = 'ag-Grid-SelectionColumn'
+const SELECTION_COLUMN_WIDTH_PX = 48
+
+type StaticSelectionColumnClass = string | string[]
+type StaticSelectionColumnStyle = Record<string, string | number>
+
+type SelectionColumnHostOptions<Row extends Record<string, any>> = Pick<ColDef<Row>,
+  | 'flex'
+  | 'headerName'
+  | 'headerTooltip'
+  | 'initialFlex'
+  | 'initialWidth'
+  | 'maxWidth'
+  | 'minWidth'
+  | 'suppressAutoSize'
+  | 'suppressSizeToFit'
+  | 'width'
+> & {
+  cellClass?: StaticSelectionColumnClass
+  cellStyle?: StaticSelectionColumnStyle
+  headerClass?: StaticSelectionColumnClass
+  headerStyle?: StaticSelectionColumnStyle
+}
+
+const createSelectionColumnOptions = <Row extends Record<string, any>>(
+  selectionColumnDef: TableDataGridGridOptions<Row>['selectionColumnDef'] = {},
+): Partial<SelectionColumnHostOptions<Row>> => {
+  const {
+    cellClass,
+    cellStyle,
+    flex,
+    headerClass,
+    headerName,
+    headerStyle,
+    headerTooltip,
+    initialFlex,
+    initialWidth,
+    maxWidth,
+    minWidth,
+    suppressAutoSize,
+    suppressSizeToFit,
+    width,
+  } = selectionColumnDef
+
+  const options: Partial<SelectionColumnHostOptions<Row>> = {
+    flex,
+    headerName,
+    headerTooltip,
+    initialFlex,
+    initialWidth,
+    maxWidth,
+    minWidth,
+    suppressAutoSize,
+    suppressSizeToFit,
+    width,
+  }
+
+  if (cellClass && typeof cellClass !== 'function') {
+    options.cellClass = cellClass as SelectionColumnHostOptions<Row>['cellClass']
+  }
+  if (cellStyle && typeof cellStyle !== 'function') {
+    options.cellStyle = cellStyle as SelectionColumnHostOptions<Row>['cellStyle']
+  }
+  if (headerClass && typeof headerClass !== 'function') {
+    options.headerClass = headerClass as SelectionColumnHostOptions<Row>['headerClass']
+  }
+  if (headerStyle && typeof headerStyle !== 'function') {
+    options.headerStyle = headerStyle as SelectionColumnHostOptions<Row>['headerStyle']
+  }
+
+  return Object.fromEntries(
+    Object.entries(options).filter(([, value]) => value !== undefined),
+  ) as Partial<SelectionColumnHostOptions<Row>>
+}
+
+const createSelectionColumnDef = <Row extends Record<string, any>>({
+  agGridOptions,
+  rowSelection,
+}: {
+  agGridOptions: TableDataGridGridOptions<Row>
+  rowSelection: TableDataGridRowSelectionMode
+}): ColDef<Row> | undefined => {
+  if (rowSelection !== 'multiple') {
+    return undefined
+  }
+
+  const selectionColumnOptions = createSelectionColumnOptions(agGridOptions.selectionColumnDef)
+  const width = selectionColumnOptions.width ?? SELECTION_COLUMN_WIDTH_PX
+  const minWidth = selectionColumnOptions.minWidth ?? SELECTION_COLUMN_WIDTH_PX
+  const maxWidth = selectionColumnOptions.maxWidth ?? Math.max(width, minWidth, SELECTION_COLUMN_WIDTH_PX)
+
+  return {
+    ...selectionColumnOptions,
+    cellRenderer: TableDataGridSelectionCell,
+    colId: SELECTION_COLUMN_ID,
+    headerComponent: TableDataGridSelectionHeader,
+    lockPosition: 'left',
+    maxWidth,
+    minWidth,
+    pinned: 'left',
+    resizable: false,
+    sortable: false,
+    suppressHeaderMenuButton: true,
+    suppressMovable: true,
+    width,
+  }
+}
 
 export const useDatatableColumnDefs = <Row extends Record<string, any>>({
   config,
   grid,
-  selection,
   slots,
 }: {
   config: TableDataGridColumnDefsConfig<Row>
   grid: TableDataGridColumnDefsGrid
-  selection: TableDataGridColumnDefsSelection<Row>
   slots: TableDataGridColumnDefsSlots
 }) => {
   const {
+    agGridOptions,
     cellAttrs,
     headers,
     resolvedTableConfig,
+    rowSelection,
   } = config
   const {
     displayedColumnIndexesByKey,
   } = grid
-  const {
-    selectionColumnDef,
-  } = selection
   const {
     slots: componentSlots,
   } = slots
@@ -78,6 +188,11 @@ export const useDatatableColumnDefs = <Row extends Record<string, any>>({
     cellRenderer: TableDataGridCellRenderer,
     ...header.agGridColumnOptions,
   })
+
+  const selectionColumnDef = computed<ColDef<Row> | undefined>(() => createSelectionColumnDef({
+    agGridOptions: agGridOptions.value,
+    rowSelection: rowSelection.value,
+  }))
 
   const columnDefs = computed<Array<ColDef<Row>>>(() => {
     const tableColumnDefs = orderedHeaders.value.map(createColumnDef)
